- Code cleanup crew.
git-svn-id: https://openxcom.svn.sourceforge.net/svnroot/openxcom/trunk@179 11c4a5ed-7179-4546-81a6-367b3be9d812

+++ src/Interface/Window.cpp
@@ -145,7 +145,7 @@ void Window::draw()
 {
 	SDL_Rect cropper;
 	SDL_Rect square;
-	int color = _color;
+	Uint8 color = _color;
 	
 	clear();
